requirements.txt
APScheduler==3.6.3
astroid==2.4.2
certifi==2020.11.8
cffi==1.14.4
chardet==3.0.4
colorama==0.4.4
cryptography==3.3.1
decorator==4.4.2
idna==2.10
iqoptionapi==6.8.9.1
isort==5.6.4
lazy-object-proxy==1.4.3
mccabe==0.6.1
pycparser==2.20
pylint==2.6.0
python-dateutil==2.8.1
python-telegram-bot==13.1
pytz==2020.4
requests==2.25.0
six==1.15.0
toml==0.10.2
tornado==6.1
tzlocal==2.1
urllib3==1.26.2
websocket-client==0.56.0
wrapt==1.12.1
